Add support for domain and local parts
From draft-09 of the WKD spech the client should include the domain and
local part of the email address in the request. Let's add support for
that.

If the domain is not provided in the url path we parse it from the
requested domain. So we add multi-domain support even if the advanced
mode of WKD is not being used.

Closes: #4